+++ state/src/call_package/mod.rs
@@ -1,0 +1,1 @@
+pub mod contract;
